+++ highwind-meta-box-color.js
@@ -1,0 +1,7 @@
+/**
+ * Highwind Light Color Picker
+ */
+jQuery(document).ready(function($){
+	$('.highwind-header-bg-color').wpColorPicker();
+	$('.highwind-header-title-color').wpColorPicker();
+});