+++ changelog.md
@@ -1,5 +1,10 @@
 # Highwind Light changelog
 
+## 1.1.1
+* ADDED   - Color picker for the Header. Now you can change the color of the header and title per post or page.
+* ADDED   - Localization for United Kingdom (en_GB)
+* UPDATED - POT file.
+
 ## 1.0.1
 * ADDED - screenshot.png
 * ADDED - Page support for excerpts.